move ToDitest impls to corresponding files
